project files moved to other directories
