+++ AppVoor/backend_scripts/parameter_search.py
@@ -7,7 +7,7 @@ from sklearn.model_selection import GridSearchCV
 from skopt import BayesSearchCV
 from skopt.space import Real, Integer, Categorical
 
-from switcher import Switch
+from backend_scripts.switcher import Switch
 
 NpArray = np.ndarray
 DataFrame = pd.DataFrame
