+++ AppVoor/backend_scripts/score.py
