+++ AppVoor/frontend_scripts/modified_widgets.py
@@ -1,4 +1,4 @@
-from PyQt5.QtCore import Qt, pyqtSlot, pyqtSignal
+from PyQt5.QtCore import Qt, pyqtSignal
 from PyQt5.QtWidgets import QPushButton, QFileDialog
