+++ AppVoor/frontend_scripts/pop_up.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 
 from PyQt5.QtGui import QFont
-from PyQt5.QtWidgets import QMessageBox, QPushButton
+from PyQt5.QtWidgets import QMessageBox
 
 
 class PopUp(ABC):
