+++ AppVoor/jsonInfo/help.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from is_data import DataEnsurer
+from backend_scripts.is_data import DataEnsurer
 from jsonInfo.json_to_data import JSONMessage
