+++ AppVoor/run.py
@@ -7,21 +7,21 @@ from PyQt5.QtCore import QRect, QThreadPool, QThread
 from PyQt5.QtGui import QFont, QTextCursor
 from PyQt5.QtWidgets import QApplication
 
-from auto_ml import JarAutoML, AutoExecutioner
-from estimator_creation import EstimatorCreator
-from feature_selection import FeatureSelectorCreator
-from global_vars import GlobalVariables
+from backend_scripts.auto_ml import JarAutoML, AutoExecutioner
+from backend_scripts.estimator_creation import EstimatorCreator
+from backend_scripts.feature_selection import FeatureSelectorCreator
+from backend_scripts.global_vars import GlobalVariables
 from jsonInfo.welcome import WelcomeMessenger
-from load_data import LoaderCreator
-from model_creation import SBSModelCreator
-from modified_widgets import QDragAndDropButton, QLoadButton
-from parallel import LongWorker, EmittingStream
-from parameter_search import ParameterSearchCreator
-from parameter_search import BayesianSearchParametersPossibilities, GridSearchParametersPossibilities
-from pop_up import PopUp, WarningPopUp, CriticalPopUp
-from result_creation import FCreator, SBSResult
-from switcher import Switch
-from view import Window
+from backend_scripts.load_data import LoaderCreator
+from backend_scripts.model_creation import SBSModelCreator
+from frontend_scripts.modified_widgets import QDragAndDropButton, QLoadButton
+from frontend_scripts.parallel import LongWorker, EmittingStream
+from backend_scripts.parameter_search import ParameterSearchCreator
+from backend_scripts.parameter_search import BayesianSearchParametersPossibilities, GridSearchParametersPossibilities
+from frontend_scripts.pop_up import PopUp, WarningPopUp, CriticalPopUp
+from backend_scripts.result_creation import FCreator, SBSResult
+from backend_scripts.switcher import Switch
+from frontend_scripts.view import Window
 from forms import resources
 
 DataFrame = pd.DataFrame
@@ -96,13 +96,13 @@ class EstimatorParametersPossibilities(Switch):
 
 class HomeWindow(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_next.clicked.connect(self.next)
         self.on_load()
 
     def on_load(self):
-        messenger = WelcomeMessenger(file_path=".\\jsonInfo\\welcomeMessage.json")
+        messenger = WelcomeMessenger(file_path="./jsonInfo/welcomeMessage.json")
         text = str(messenger)
         self.lbl_description.setText(text)
 
@@ -122,8 +122,8 @@ class HomeWindow(Window):
 
 class DataSetWindow(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.file_type: str = "CSV"
         self.last: str = "any"
         self.critical_pop_up: PopUp = CriticalPopUp()
@@ -276,8 +276,8 @@ class DataSetWindow(Window):
 
 class MLTypeWindow(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_automl.clicked.connect(lambda: self.useful_info_pop_up("auto_machine_learning"))
@@ -318,8 +318,8 @@ class MLTypeWindow(Window):
 
 class AutoLoad(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         sys.stdout = EmittingStream(textWritten=self.add_info)
 
         self.lbl_cancel.mouseReleaseEvent = self.cancel_training
@@ -360,12 +360,13 @@ class AutoLoad(Window):
     def train_model(self) -> None:
         automl_ml = JarAutoML(10, False, 5000)
         model = AutoExecutioner(automl_ml)
-        self.add_info(str(model)+"\n\n")
+        self.add_info(str(model) + "\n\n")
         data_frame = global_var.data_frame
         model.train_model(data_frame)
 
     def cancel_training(self, event) -> None:
         """Show a Warning pop up and then if user wants to finished the app, close it"""
+
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Cancelar entrenamiento"
@@ -382,6 +383,7 @@ class AutoLoad(Window):
 
     def handle_error(self, error) -> None:
         """Print error message to the QTextEdit"""
+
         def write_error():
             for i in info:
                 self.add_info(i)
@@ -423,8 +425,8 @@ class AutoLoad(Window):
 
 class StepByStepLoad(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         sys.stdout = EmittingStream(textWritten=self.add_info)
 
         self.lbl_cancel.mouseReleaseEvent = self.cancel_training
@@ -568,8 +570,8 @@ class StepByStepLoad(Window):
 
 class PredictionType(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_Classification.clicked.connect(lambda: self.useful_info_pop_up("classification"))
@@ -597,8 +599,8 @@ class PredictionType(Window):
 
 class ClassificationSelection(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_KNN.clicked.connect(lambda: self.useful_info_pop_up("knn"))
@@ -629,8 +631,8 @@ class ClassificationSelection(Window):
 
 class RegressionSelection(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_Lasso.clicked.connect(lambda: self.useful_info_pop_up("lasso"))
@@ -661,8 +663,8 @@ class RegressionSelection(Window):
 
 class ClusteringSelection(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_Affinity_Propagation.clicked.connect(lambda: self.useful_info_pop_up("affinity_propagation"))
@@ -693,8 +695,8 @@ class ClusteringSelection(Window):
 
 class WantFeatureSelection(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_FSM.clicked.connect(lambda: self.useful_info_pop_up("feature_selection"))
@@ -726,8 +728,8 @@ class WantFeatureSelection(Window):
 
 class FeatureSelectionMethod(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_FS.clicked.connect(lambda: self.useful_info_pop_up("forward_feature_selection"))
@@ -754,8 +756,8 @@ class FeatureSelectionMethod(Window):
 
 class WantHyperparameterSearch(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_Search_Hiperparameters.clicked.connect(lambda: self.useful_info_pop_up("parameter_search"))
@@ -789,8 +791,8 @@ class WantHyperparameterSearch(Window):
 
 class HyperparameterMethod(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_Bayesian_Search.clicked.connect(lambda: self.handle_input("Bayesian"))
@@ -840,8 +842,8 @@ class HyperparameterMethod(Window):
 
 class FinalResult(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.lbl_home.mouseReleaseEvent = self.back
 
     def back(self, event) -> None:
@@ -855,8 +857,8 @@ class FinalResult(Window):
 
 class AffinityPropagationParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_convergencia.clicked.connect(lambda: self.useful_info_pop_up("APROPAGATION_convergencia"))
@@ -867,7 +869,6 @@ class AffinityPropagationParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -898,8 +899,8 @@ class AffinityPropagationParameters(Window):
 
 class GaussianNBParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_variable_refinamiento.clicked.connect(lambda: self.useful_info_pop_up("GNB_refinamiento"))
@@ -907,7 +908,6 @@ class GaussianNBParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -935,8 +935,8 @@ class GaussianNBParameters(Window):
 
 class KMeansParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_clusters.clicked.connect(lambda: self.useful_info_pop_up("KMEANS_n_clusters"))
@@ -947,7 +947,6 @@ class KMeansParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -979,8 +978,8 @@ class KMeansParameters(Window):
 
 class KNeighborsClassifierParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_numero_vecinos.clicked.connect(lambda: self.useful_info_pop_up("KNN_n_vecinos"))
@@ -991,7 +990,6 @@ class KNeighborsClassifierParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1023,8 +1021,8 @@ class KNeighborsClassifierParameters(Window):
 
 class LassoParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_alfa.clicked.connect(lambda: self.useful_info_pop_up("LASSO_alfa"))
@@ -1035,7 +1033,6 @@ class LassoParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1067,8 +1064,8 @@ class LassoParameters(Window):
 
 class LinearSVCParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_parametro_regularizacion.clicked.connect(lambda: self.useful_info_pop_up("LSVC_C"))
@@ -1079,7 +1076,6 @@ class LinearSVCParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1111,8 +1107,8 @@ class LinearSVCParameters(Window):
 
 class LinearSVRParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_parametro_regularizacion.clicked.connect(lambda: self.useful_info_pop_up("LSVR_C"))
@@ -1123,7 +1119,6 @@ class LinearSVRParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1155,8 +1150,8 @@ class LinearSVRParameters(Window):
 
 class MeanShiftParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_ancho_banda.clicked.connect(lambda: self.useful_info_pop_up("MEANSHIFT_ancho_banda"))
@@ -1169,7 +1164,6 @@ class MeanShiftParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1201,8 +1195,8 @@ class MeanShiftParameters(Window):
 
 class MiniBatchKMeansParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_clusters.clicked.connect(lambda: self.useful_info_pop_up("MINIKMEANS_n_clusters"))
@@ -1213,7 +1207,6 @@ class MiniBatchKMeansParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1245,8 +1238,8 @@ class MiniBatchKMeansParameters(Window):
 
 class SGDClassifierParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_alfa.clicked.connect(lambda: self.useful_info_pop_up("SGD_alfa"))
@@ -1257,7 +1250,6 @@ class SGDClassifierParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1289,8 +1281,8 @@ class SGDClassifierParameters(Window):
 
 class SVCParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_parametro_regularizacion.clicked.connect(lambda: self.useful_info_pop_up("SVC_C"))
@@ -1301,7 +1293,6 @@ class SVCParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
@@ -1333,8 +1324,8 @@ class SVCParameters(Window):
 
 class SVRParameters(Window):
 
-    def __init__(self, window: str, help_message_path: str = ".\\jsonInfo\\helpMessage.json") -> None:
-        super().__init__(window, help_message_path)
+    def __init__(self, window: str) -> None:
+        super().__init__(window)
         self.btn_back.clicked.connect(self.back)
 
         self.btn_info_parametro_regularizacion.clicked.connect(lambda: self.useful_info_pop_up("SVR_C"))
@@ -1345,7 +1336,6 @@ class SVRParameters(Window):
         self.btn_next.clicked.connect(self.next)
 
     def next(self):
-
         def show_last_warning() -> bool:
             pop_up: PopUp = WarningPopUp()
             title = "Listo para entrenar"
