+++ AppVoor/tests/creator_test.py
@@ -1,10 +1,10 @@
 import unittest
 
-from estimator_creation import EstimatorCreator
-from feature_selection import FeatureSelectorCreator
-from load_data import LoaderCreator
-from model_creation import SBSModelCreator
-from parameter_search import ParameterSearchCreator
+from backend_scripts.estimator_creation import EstimatorCreator
+from backend_scripts.feature_selection import FeatureSelectorCreator
+from backend_scripts.load_data import LoaderCreator
+from backend_scripts.model_creation import SBSModelCreator
+from backend_scripts.parameter_search import ParameterSearchCreator
 
 
 class MyTestCase(unittest.TestCase):
